\name{this.path-defunct}
\alias{this.path-defunct}


\alias{ArgumentParser}
\alias{ArgumentParser-class}
\alias{this.path_ArgumentParser-class}


\alias{add.argument}
\alias{add.argument,ArgumentParser-method}
\alias{add.argument,this.path_ArgumentParser-method}


\alias{add.help}
\alias{add.skip}
\alias{add.version}
\alias{add.help,ArgumentParser-method}
\alias{add.skip,ArgumentParser-method}
\alias{add.version,ArgumentParser-method}
\alias{add.help,this.path_ArgumentParser-method}
\alias{add.skip,this.path_ArgumentParser-method}
\alias{add.version,this.path_ArgumentParser-method}


\alias{add.subparsers}
\alias{add.subparsers,ArgumentParser-method}
\alias{add.subparsers,this.path_ArgumentParser-method}


\alias{Args}


\alias{Commands}


\alias{dedent}


\alias{delayedAssign2}


\alias{file.open}


\alias{list.files2}
\alias{dir2}


\alias{Missing}


\alias{normalizeAgainst}


\alias{commandPrompt}


\alias{commandQuote}
\alias{shEncode}


\alias{path.contract}


\alias{pseudoglobalenv}


\alias{python}


\alias{R}
\alias{Rcmd}
\alias{Rscript}
\alias{Rterm}


\alias{readArgs}
\alias{writeArgs}


\alias{has.ext}
\alias{scan2}
\alias{format4scan}
\alias{setReadWriteArgsMethod}


\alias{withArgs}


\title{
Defunct Functions in Package \strong{this.path}
}
\description{
The functions or variables listed here are no longer part of \code{this.path}
as they are no longer needed or have been moved to a new package.
}
\usage{
ArgumentParser(...)

\S4method{add.argument}{this.path_ArgumentParser}(...)

\S4method{add.help}{this.path_ArgumentParser}(...)

\S4method{add.skip}{this.path_ArgumentParser}(...)

\S4method{add.version}{this.path_ArgumentParser}(...)

\S4method{add.subparsers}{this.path_ArgumentParser}(...)

Args(...)

Commands(...)

dedent(...)

delayedAssign2(...)

file.open(...)

list.files2(...)
dir2(...)

Missing(...)

normalizeAgainst(...)

commandPrompt(...)

commandQuote(...)
shEncode(...)

path.contract(...)

pseudoglobalenv(...)

python(...)

R(...)
Rcmd(...)
Rscript(...)
Rterm(...)

writeArgs(...)
readArgs(...)

has.ext(...)
scan2(...)
format4scan(...)
setReadWriteArgsMethod(...)

withArgs(...)
}
\arguments{
  \item{\dots}{
  }
}
\seealso{
\code{\link[base]{Defunct}}
}
