\name{this.path-deprecated}
\alias{this.path-deprecated}
\alias{this.path2}
\alias{this.dir2}
\alias{this.dir3}
\title{
Deprecated Functions in Package \pkg{this.path}
}
\description{
These functions are provided for compatibility with older versions of \R only,
and may be defunct as soon as the next release.
}
\usage{
this.path2(...)
this.dir2(...)

this.dir3(...)
}
\arguments{
  \item{\dots}{
arguments passed to \code{\link{this.path}}.
  }
}
\seealso{
\code{\link{this.path2-deprecated}}\verb{
}\code{\link{this.dir2-deprecated}}\verb{
}\code{\link{this.dir3-deprecated}}
}
