\name{source.exprs}
\alias{source.exprs}
\title{Evaluating and Print Expressions}
\description{
  \code{source.exprs()} evaluates and auto-prints expression as if in a
  toplevel contenxt.
}
\usage{
source.exprs(exprs, evaluated = FALSE, envir = parent.frame(),
             echo = TRUE, print.eval = TRUE)
}
\arguments{
  \item{exprs, evaluated, envir, echo, print.eval}{See
    \codelink3{base:source}{withAutoprint}{()}.}
}
