\name{this.path-defunct}
\alias{this.path-defunct}
\alias{this.path2}
\alias{this.dir2}
\alias{this.dir3}
\title{
Defunct Functions in Package \pkg{this.path}
}
\description{
The functions or variables listed here are no longer part of \pkg{this.path} as
they are no longer needed.
}
\usage{
# Defunct in 1.1.0
this.path2(...)  # use 'this.path(..., default = NULL)' instead
this.dir2(...)   # use 'this.dir(..., default = NULL)' instead
this.dir3(...)   # use 'this.dir(..., default = getwd())' instead
}
\arguments{
  \item{\dots}{
  }
}
\seealso{
\code{\link{this.path}}, \code{\link{this.dir}}
}
