\name{this.path-package}
\alias{this.path-package}
\docType{package}
\title{
\packageTitle{this.path}
}
\description{
\packageDescription{this.path}
}
\details{
There are only two functions from this package; this.path and this.dir

this.path() returns the full path of the executing script.

this.dir() is a shorter way of writing dirname(this.path()).
It returns the full path of the directory where the executing script is located.
}
\author{
\packageAuthor{this.path}

Maintainer: \packageMaintainer{this.path}
}
\keyword{ package }
\seealso{
\code{\link[this.path]{this.path}}

\code{\link[base]{source}}

\code{\link[base]{sys.source}}

\code{\link[utils]{Rscript}}
}
